PROD-260 Create 객체 dispatch 경계를 분리한다
