+++ packages/fedify/src/federation.ts
@@ -1,7 +1,7 @@
 import { createFederation, MemoryKvStore } from '@fedify/fedify';
 import { Create, Follow, Undo } from '@fedify/vocab';
 import { resolveConfiguredLocalInstance } from '@kosmo/core/local-instance';
-import { handleInboundCreateNote } from './inbound-create-note';
+import { handleInboundCreate } from './inbound-create';
 import { handleInboundFollow, handleInboundUndo } from './inbound-follow';
 import { ensureDrizzleLocalProfileActor } from './local-actor-store';
 import { createLocalProfilePerson } from './local-profile-person';
@@ -61,7 +61,7 @@ federation
 federation
   .setInboxListeners('/ap/actor/{identifier}/inbox', '/inbox')
   .on(Create, async (context, create) => {
-    await handleInboundCreateNote(context, create);
+    await handleInboundCreate(context, create);
   })
   .on(Follow, handleInboundFollow)
   .on(Undo, handleInboundUndo);
