+++ packages/fedify/src/inbound-create.ts
@@ -1,0 +1,54 @@
+import '@kosmo/core/polyfill';
+
+import { Note } from '@fedify/vocab';
+import { InstanceState } from '@kosmo/core/enums';
+import { handleInboundCreateNote } from './inbound-create-note';
+import { findStoredRemoteProfileActorByUri } from './remote-actor-materialization';
+import type { InboxContext } from '@fedify/fedify';
+import type { Create } from '@fedify/vocab';
+import type { InboundCreateNoteMaterializationInput } from './inbound-create-note';
+
+export type InboundCreateMaterializationInput = InboundCreateNoteMaterializationInput;
+
+const getNow = () => Temporal.Now.instant();
+
+const uniqueHref = (uris: URL[]): string | undefined => {
+  const hrefs = new Set(uris.map((uri) => uri.href));
+
+  return hrefs.size === 1 ? hrefs.values().next().value : undefined;
+};
+
+export const handleInboundCreate = async (
+  context: InboxContext<void>,
+  create: Create,
+  receivedAt: Temporal.Instant = getNow(),
+): Promise<InboundCreateMaterializationInput | undefined> => {
+  const actorUri = uniqueHref(create.actorIds);
+  const objectUri = uniqueHref(create.objectIds);
+
+  if (!actorUri || !objectUri) {
+    return undefined;
+  }
+
+  const storedActor = await findStoredRemoteProfileActorByUri(actorUri);
+  if (
+    !storedActor ||
+    (storedActor.instance.state !== InstanceState.ACTIVE &&
+      storedActor.instance.state !== InstanceState.UNRESPONSIVE)
+  ) {
+    return undefined;
+  }
+
+  let object;
+  try {
+    object = await create.getObject({ documentLoader: context.documentLoader });
+  } catch {
+    return undefined;
+  }
+
+  if (object instanceof Note) {
+    return handleInboundCreateNote({ actorUri, note: object, objectUri, receivedAt });
+  }
+
+  return undefined;
+};
